requirements.txt
grpcio==1.76.0
protobuf>=6.33.0
pydantic>=2.12.3